==> exercises/assignment/1_data_pipeline/1_clean_data.ipynb <==
# Clean retail data using Python

import pandas as pd

df = pd.read_csv("1_initial_train.csv")

print("\n Dataset Info :")
print(df.info())

df.head()

df["Ship Mode"].value_counts()

df["Country"].value_counts()

df["Region"].value_counts()

df["Segment"].value_counts()

### Handling missing values

print("\n Missing Values:")
print(df.isnull().sum())

There are 11 missing values in the Postal Code column.

print("Filling missing Postal Code with 'N/A' (convert to string)")
df['Postal Code'] = df['Postal Code'].fillna('N/A').astype(str)

print("\n Missing Values after cleaning:")
print(df.isnull().sum())

df.info()

### Handling duplicate rows

print(f"\n Duplicate Rows Found: {df.duplicated().sum()}")

There are no duplicates in this dataset.

### Handling data types

print("\n Data Type Issues:")
print(df.dtypes)

print("Converting Date Columns")
df['Order Date'] = pd.to_datetime(df['Order Date'], format='%d/%m/%Y', errors='coerce')  
df['Ship Date'] = pd.to_datetime(df['Ship Date'], format='%d/%m/%Y', errors='coerce')  

print("\Converting Essential Columns to 'Category' Type")
category_columns = ['Ship Mode', 'Segment','Category', 'Sub-Category']
df[category_columns] = df[category_columns].astype('category')

### Final dataset after cleaning

#sorting data by order date
df.sort_values(by=['Order Date'], inplace=True, ascending=True)

print("\n Final Dataset After Cleaning:")
print(df.info())


df.head(3)

df.to_csv("data/2_cleaned_data.csv", index=False)

# Conclusion

In this section, the retail dataset was successfully cleaned using Python. The following steps were performed:

* Data Inspection: The dataset's structure, including column names, data types, and basic statistics, was examined.

* Handling Missing Values: Missing values in the "Postal Code" column were identified and filled, converting them to 'N/A' as strings.

* Handling Duplicates: The dataset was checked for duplicate rows, and it was confirmed that there were no duplicates.

* Handling Data Types: The "Order Date" and "Ship Date" columns were converted to datetime format to ensure proper handling of date-related operations.

* Sorting Data: The dataset was sorted by "Order Date" to maintain a logical order for subsequent analysis.

The cleaned dataset is now ready for processing sales history with PySpark.
