==> exercises/assignment/1_data_pipeline/2_process_sales_history.ipynb <==
# Process sales history with PySpark  

from pyspark.sql import SparkSession
from pyspark.sql.functions import col, sum, avg, count, year, month, round, regexp_replace

# Start Spark session
spark = SparkSession.builder \
    .appName("Sales History Processing") \
    .getOrCreate()

- Load the cleaned dataset

df = spark.read.csv("dbfs:/FileStore/Retail_Store_Chain_Analysis/2_cleaned_data.csv", header=True, inferSchema=True, multiLine=True, escape='"')


# Checking missing values for all columns
null_count = df.select([sum(col(c).isNull().cast("int")).alias(c) for c in df.columns])
display(null_count)

df.printSchema()

# Round Sales column to 3 decimal places
df = df.withColumn("Sales", round(col("Sales"), 3))
df.select('Sales').show(truncate=False)


df.isEmpty()

df.count()

df.select('Row ID').distinct().count()

df.select('Order ID').distinct().count()

df.select('Customer ID').distinct().count()

# Extract Year and Month
df = df.withColumn('Order Year', year('Order Date'))
df = df.withColumn('Order Month', month('Order Date'))

df.select('Order Year').distinct().show()

# Aggregation: Sales Summary by Year
sales_by_year = df.groupBy('Order Year').agg(
    sum("Sales").alias("Total Sales"),
    avg("Sales").alias("Average Sales"),
    count("Order ID").alias("Total Orders")
)
sales_by_year.show()

Sales peaked in 2018, with the highest total revenue and orders. 2016 had the highest average sales per order. Steady growth is observed over the years.

# Aggregate sales by Year and Month
seasonal_sales = df.groupBy('Order Year', 'Order Month').agg(
    sum('Sales').alias('Total Sales'),
    avg('Sales').alias('Average Sales'),
    count('Order ID').alias('Total Orders')
)
seasonal_sales.show()


df.columns

df.printSchema()

df = df.withColumn("Postal Code", regexp_replace("Postal Code", "\\.0$", ""))

#Save the DataFrame as a CSV file
df.write.option("header", "true").csv("dbfs:/FileStore/Retail_Store_Chain_Analysis/3_processed_data.csv")

# Save the processed data to a Parquet file for future analysis
df.write.partitionBy("Region", "Category", "Order Year").parquet("dbfs:/FileStore/Retail_Store_Chain_Analysis/3_processed_data.parquet", mode="overwrite")

###  Conclusion

The sales history data was processed using PySpark. The following steps were performed:

* Loaded the cleaned dataset.

* Checked for missing values and rounded the "Sales" column.

* Extracted year and month from the "Order Date" column.

* Aggregated sales by year and by year-month.

* Cleaned the "Postal Code" column.

The processed dataset is saved as both a CSV and a Parquet file for future analysis.
