==> exercises/assignment/1_data_pipeline/4_pyspark_to_sql.ipynb <==
# Inserting Datas Into Database Tables

from pyspark.sql import SparkSession
from sqlalchemy import create_engine
from pyspark.sql.functions import when, col
import yaml
import os

# Get the absolute path to the root directory
root_dir = os.path.abspath(os.path.join(os.getcwd(), os.pardir))

# Construct the path to config.yaml
config_path = os.path.join(root_dir, 'config.yaml')

# Load configurations from config.yaml
with open(config_path, 'r') as file:
    config = yaml.safe_load(file)

# Database credentials
db_username = config['database']['username']
db_password = config['database']['password']
db_host = config['database']['host']
db_port = config['database']['port']
db_name = config['database']['dbname']

# PySpark settings
input_file = config['pyspark']['input_file']

spark = SparkSession.builder.appName("InsertDataIntoPostgreSQL").getOrCreate()

# Read input CSV file
df = spark.read.csv(input_file, header=True, inferSchema=True, multiLine=True, escape='"')

# df=spark.read.csv("export (1).csv",header=True, inferSchema=True, multiLine=True, escape='"')

# # PostgreSQL database connection details
# db_name = "retail_store_db"
# db_host = "localhost"
# db_username = "retail_db_user"
# db_password = "retail_db_password"
# db_port = "5432"

# Create SQLAlchemy engine for PostgreSQL and establish connection to database
engine = create_engine(f'postgresql+psycopg2://{db_username}:{db_password}@{db_host}:{db_port}/{db_name}')
conn = engine.connect()

### Data Insertion Overview

The following tables need to be populated as part of the data pipeline:

- **Sales**: Contains transactional sales data, including product details, customer information, and order metadata.
- **Orders**: Stores information about each order placed by customers.
- **Customers**: Maintains customer details.
- **Geography**: Holds postal code mappings to geographical locations.
- **Products**: Contains product-related details.


 **Customer Table**
<!-- CREATE TABLE Customers (
    CustomerID VARCHAR PRIMARY KEY,
    CustomerName VARCHAR NOT NULL,
    Segment VARCHAR CHECK (Segment IN ('Consumer', 'Corporate', 'Home Office'))
); -->

## Customers Table
customer_df = df.select("Customer ID", "Customer Name", "Segment").distinct()
customer_pd = customer_df.toPandas()
customer_pd.columns = ["customerid", "customername", "segment"]
customer_pd.to_sql("customers", con=conn, if_exists='append', index=False)

customer_pd

**Geography Dimension Table**
<!-- CREATE TABLE Geography (
    PostalCode VARCHAR PRIMARY KEY,
    City VARCHAR NOT NULL,
    State VARCHAR NOT NULL,
    Region VARCHAR CHECK (Region IN ('West', 'East', 'Central', 'South'))
);
 -->

# Check 'N/A' rows in df
df.select("Postal Code", "City", "State", "Region").filter(col("Postal Code") == "N/A").show()

df = df.withColumn("Postal Code", when(col("Postal Code") == "N/A", "5401").otherwise(col("Postal Code")))
df.select("Postal Code", "City", "State", "Region").filter(col("Postal Code") == "N/A").show()


Since all missing postal codes are from **Burlington, Vermont (East region)**, they are being replaced with the correct postal code **"05401"** .

## Geography Table
geography_df = df.select("Postal Code", "City", "State", "Region").distinct()
geography_pd = geography_df.toPandas()
geography_pd.columns = ["postalcode", "city", "state", "region"]

# Identify duplicate postal codes
duplicate_postal_codes = geography_pd[geography_pd.duplicated(subset=['postalcode'], keep=False)]
print(duplicate_postal_codes)


df = df.withColumn("Postal Code", when(df["City"] == "San Diego", 92101).otherwise(df["Postal Code"]))

The dataset contains duplicate postal codes for different cities. Specifically, **92024** is assigned to both **San Diego** and **Encinitas** in California. To resolve this inconsistency, replaced the **Postal Code** for **San Diego** with **92101**.

## Geography Table
geography_df = df.select("Postal Code", "City", "State", "Region").distinct()
geography_pd = geography_df.toPandas()
geography_pd.columns = ["postalcode", "city", "state", "region"]
geography_pd.to_sql("geography", con=conn, if_exists='append', index=False)


 **Product Dimension Table**
<!-- CREATE TABLE Products (
    ProductID VARCHAR NOT NULL,
    ProductName VARCHAR NOT NULL,
    SubCategory VARCHAR NOT NULL,
    Category VARCHAR CHECK (Category IN ('Furniture', 'Office Supplies', 'Technology')),
    PRIMARY KEY (ProductID,ProductName)
); -->

## Products Table
product_df = df.select("Product ID", "Product Name", "Sub-Category", "Category").distinct()
product_pd = product_df.toPandas()
product_pd.columns = ["productid", "productname", "subcategory", "category"]

# Check for duplicate values based on "productid"
duplicate_rows = product_pd[product_pd.duplicated(subset=["productid"], keep=False)]
# Print duplicate values
print(duplicate_rows)


In the Products table, since ProductID alone is not unique, a composite primary key (ProductID, ProductName) is used.

product_pd.to_sql("products", con=conn, if_exists='append', index=False)

**Order Dimension Table**
<!-- CREATE TABLE Orders (
    OrderID VARCHAR PRIMARY KEY,
    OrderDate DATE NOT NULL,
    ShipDate DATE NOT NULL,
    ShipMode VARCHAR CHECK (ShipMode IN ('First Class', 'Second Class', 'Standard Class', 'Same Day'))
); -->

## Orders Table
order_df = df.select("Order ID", "Order Date", "Ship Date", "Ship Mode").distinct()
order_pd = order_df.toPandas()
order_pd.columns = ["orderid", "orderdate", "shipdate", "shipmode"]
order_pd.to_sql("orders", con=conn, if_exists='append', index=False)

**Fact Table: Sales**
<!-- CREATE TABLE Retail_Sales (
    RowID SERIAL PRIMARY KEY,
    OrderID VARCHAR NOT NULL,
    CustomerID VARCHAR NOT NULL,
    PostalCode VARCHAR NOT NULL,
    ProductID VARCHAR NOT NULL,
    ProductName VARCHAR NOT NULL,
    Sales DECIMAL(10,2) NOT NULL CHECK (Sales >= 0), 
    OrderYear INT CHECK (OrderYear BETWEEN 2000 AND 2100),
    OrderMonth INT CHECK (OrderMonth BETWEEN 1 AND 12),
    FOREIGN KEY (OrderID) REFERENCES Orders(OrderID) ON DELETE CASCADE,
    FOREIGN KEY (CustomerID) REFERENCES Customers(CustomerID) ON DELETE CASCADE,
    FOREIGN KEY (PostalCode) REFERENCES Geography(PostalCode) ON DELETE CASCADE,
    FOREIGN KEY (ProductID, ProductName) REFERENCES Products(ProductID, ProductName) ON DELETE CASCADE
); -->


retail_sales_df = df.select(
    col("Row ID").alias("rowid"),
    col("Order ID").alias("orderid"),
    col("Customer ID").alias("customerid"),
    col("Postal Code").alias("postalcode"),
    col("Product ID").alias("productid"),
    col("Product Name").alias("productname"),
    col("Sales").alias("sales"),
    col("Order Year").alias("orderyear"),
    col("Order Month").alias("ordermonth")
)

# Convert the PySpark DataFrame to Pandas DataFrame for database insertion
retail_sales_pd = retail_sales_df.toPandas()

# Insert data into the Retail_Sales table
retail_sales_pd.to_sql("retail_sales", con=engine, if_exists="append", index=False)

print("Data successfully inserted into Retail_Sales table!")


retail_sales_pd.count()

# # Insert data into PostgreSQL table
# df_pd = df.toPandas
# table_name = "main123"
# df_pd.to_sql(table_name, con=conn, if_exists='append', index=False)
# print(f"Data inserted into {table_name} successfully")

## Conclusion
Designed efficient PostgreSQL schema for retail analytics and corresponding data are then populated into the schema.
